закончил обязательную часть
