+++ backend/package.json
@@ -32,8 +32,10 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.19.2",
+    "express-winston": "^4.2.0",
     "joi": "^17.13.3",
     "mongoose": "^8.3.2",
-    "validator": "^13.11.0"
+    "validator": "^13.11.0",
+    "winston": "^3.17.0"
   }
 }
