+++ backend/src/middlewares/logger.ts
@@ -1,0 +1,20 @@
+// middlewares/logger.ts
+import winston from 'winston';
+import expressWinston from 'express-winston';
+
+const requestLogger = expressWinston.logger({
+  transports: [
+    new winston.transports.File({ filename: 'logs/request.log' }),
+  ],
+  format: winston.format.json(),
+});
+
+// логгер ошибок
+const errorLogger = expressWinston.errorLogger({
+  transports: [
+    new winston.transports.File({ filename: 'logs/error.log' }),
+  ],
+  format: winston.format.json(),
+});
+
+export { requestLogger, errorLogger };
